build updated

+++ build.sh
@@ -8,11 +8,21 @@ build_run(){
 
     cd $dir
 
-   # Clean up previous build (delete the old executable)
-    rm -f $executable
+    if [ -f "$executable" ];then
+      if [ "$source_file" -nt "$executable" ];then # check timestamp (mtime) Modification time
+        
+        # Clean up previous build (delete the old executable)
+        rm -f $executable
 
-   # Compile the scheduler.cpp file with cxxopts.hpp from the parent directory
-    g++ -I.. -o $executable $source_file
+        # Compile the scheduler.cpp file with cxxopts.hpp from the parent directory
+        g++ -I.. -o $executable $source_file
+      else 
+        echo "Executable is up-to-date. No need to recompile."
+      fi
+    else 
+      echo "First compilation"
+      gcc -I.. -o $executable $source_file
+    fi
 
    #run
     ./$executable "$@"
@@ -26,16 +36,16 @@ if [ -z "$1" ]; then
 fi
 
 case $1 in
-    "scheduler")
+      "scheduler")
        shift 1 # Remove the first argument (scheduler)
        build_run "schedulers" "scheduler.cpp" "scheduler" "$@"
-       ;;
-     "mlfq")
+      ;;
+      "mlfq")
        shift 1
        build_run "mlfq" "mlfq.cpp" "mlfq" "$@"
-       ;;
-    *)
+      ;;
+      *)
        echo "Unknown file: $1"
-       ;;
+      ;;
 esac
